Extend the Makefile to support dependencies from APT by default
The current default does not work without modifications in any scenario.

This makes it easier to build the adapter in a Debian-based system,
after simply getting SPOOLES, ARPACK, and yaml-cpp from APT.
For users that have these dependencies elsewhere, they still need
to set the dependencies themselves.

Furthermore, it sets the suggested location for CCX and the dependencies to
`$(HOME)`, a pattern often observed among new users.

This also splits each dependency to an *_INCLUDE and an *_LIBS
part, making it more flexible.